bugfix in saturate
